feat: Adapt azfa to use Cells fetch (#55)
