+++ src/configs/static-site/webpack.worker.config.ts
@@ -3,8 +3,8 @@ import { generateWorkerCommonConfig } from "../common/webpack.worker.config";
 
 const TMP_ENTRY_FILE = "./src/index.tmp.js";
 
-const generateWorkerStaticSiteConfig = (outputPath: string): Configuration  => {
-    const config: Configuration = generateWorkerCommonConfig();
+const generateWorkerStaticSiteConfig = (outputPath: string, pluginsList: any, versionId: any): Configuration  => {
+    const config: Configuration = generateWorkerCommonConfig(pluginsList);
 
     if (!config.output) config.output = {};
     config.output.path = outputPath;
@@ -16,6 +16,7 @@ const generateWorkerStaticSiteConfig = (outputPath: string): Configuration  => {
     config.plugins?.push(
         new DefinePlugin({
             PROJECT_TYPE_PATTERN_VALUE: JSON.stringify("PROJECT_TYPE:STATIC_SITE"),
+            VERSION_ID: JSON.stringify(versionId)
         })
     );
