+++ src/constants.ts
@@ -4,4 +4,4 @@ const CELLS_SITE_TEMPLATE_WORK_DIR = "azion/cells-site-template";
 export {
     CELLS_SITE_TEMPLATE_REPO,
     CELLS_SITE_TEMPLATE_WORK_DIR  
-};
+}
