+++ test/ts/configs/common/webpack.worker.config.test.ts
@@ -6,8 +6,9 @@ import { generateWorkerCommonConfig } from '../../../../dist/configs/common/webp
 
 describe('common worker webpack config', () => {
     let workerCommonConfig: Configuration;
+    const pluginsList: any = [];
     before(() => {
-        workerCommonConfig = generateWorkerCommonConfig();
+        workerCommonConfig = generateWorkerCommonConfig(pluginsList);
     });
 
     it('should export correct build base config', () => {
@@ -17,29 +18,6 @@ describe('common worker webpack config', () => {
         expect(workerCommonConfig.output?.filename).to.be.equals("function.js");
         expect(workerCommonConfig.output?.sourceMapFilename).to.be.equals("function.js.map");
         expect(workerCommonConfig.output?.globalObject).to.be.equals("this");
-    });
-
-    it('should add VirtualModulesPlugin with bootstrap module', () => {
-        const virtualModulesPlugin: any = (workerCommonConfig.plugins ?? []).filter(
-            (el: object) => el.constructor.name === "VirtualModulesPlugin"
-        )[0];
-
-        expect(virtualModulesPlugin).not.to.be.undefined;
-        expect(virtualModulesPlugin._staticModules).to.have.key('node_modules/bootstrap.js');
-
-        const bootstrapVirtualModule = virtualModulesPlugin._staticModules['node_modules/bootstrap.js'];
-        expect(bootstrapVirtualModule).to.match(/encoder: /);
-        expect(bootstrapVirtualModule).to.match(/HOST_SERVICES: /);
-        expect(bootstrapVirtualModule).to.match(/UNSIGNABLE_HEADERS: /);
-        expect(bootstrapVirtualModule).to.match(/hmac: /);
-        expect(bootstrapVirtualModule).to.match(/hash: /);
-        expect(bootstrapVirtualModule).to.match(/buf2hex: /);
-        expect(bootstrapVirtualModule).to.match(/encodeRfc3986: /);
-        expect(bootstrapVirtualModule).to.match(/guessServiceRegion: /);
-        expect(bootstrapVirtualModule).to.match(/AwsClient: /);
-        expect(bootstrapVirtualModule).to.match(/AwsV4Signer: /);
-        expect(bootstrapVirtualModule).to.match(/KV: /);
-        expect(bootstrapVirtualModule).to.match(/makeCaches: /);
     });
 
     it('should add LimitChunkCountPlugin and set max chunks to 1', () => {
